+++ dev/NetMath.js
@@ -41,18 +41,18 @@ class NetMath {
     }
     
     // Cost functions
-    static crossEntropy (target, output) {
+    static crossentropy (target, output) {
         return output.map((value, vi) => target[vi] * Math.log(value+1e-15) + ((1-target[vi]) * Math.log((1+1e-15)-value)))
                      .reduce((p,c) => p-c, 0)
     }
 
-    static meanSquaredError (calculated, desired) {
+    static meansquarederror (calculated, desired) {
         return calculated.map((output, index) => Math.pow(output - desired[index], 2))
                          .reduce((prev, curr) => prev+curr, 0) / calculated.length
     }
 
     // Weight updating functions
-    static noAdaptiveLR (value, deltaValue) {
+    static noadaptivelr (value, deltaValue) {
         return value + this.learningRate * deltaValue
     }
 
@@ -83,7 +83,7 @@ class NetMath {
                                                                                         : neuron.biasCache))
     }
 
-    static RMSProp (value, deltaValue, neuron, weightI) {
+    static rmsprop (value, deltaValue, neuron, weightI) {
 
         if(weightI!=null)
              neuron.weightsCache[weightI] = this.rmsDecay * neuron.weightsCache[weightI] + (1 - this.rmsDecay) * Math.pow(deltaValue, 2)
@@ -140,21 +140,21 @@ class NetMath {
         })
     }
 
-    static xavierNormal (size, {fanIn, fanOut}) {
+    static xaviernormal (size, {fanIn, fanOut}) {
         return fanOut || fanOut==0 ? NetMath.gaussian(size, {mean: 0, stdDeviation: Math.sqrt(2/(fanIn+fanOut))})
-                                   : NetMath.lecunNormal(size, {fanIn})
+                                   : NetMath.lecunnormal(size, {fanIn})
     }
 
-    static xavierUniform (size, {fanIn, fanOut}) {
+    static xavieruniform (size, {fanIn, fanOut}) {
         return fanOut || fanOut==0 ? NetMath.uniform(size, {limit: Math.sqrt(6/(fanIn+fanOut))})
-                                   : NetMath.lecunUniform(size, {fanIn})
+                                   : NetMath.lecununiform(size, {fanIn})
     }    
 
-    static lecunNormal (size, {fanIn}) {
+    static lecunnormal (size, {fanIn}) {
         return NetMath.gaussian(size, {mean: 0, stdDeviation: Math.sqrt(1/fanIn)})
     }
 
-    static lecunUniform (size, {fanIn}) {
+    static lecununiform (size, {fanIn}) {
         return NetMath.uniform(size, {limit: Math.sqrt(3/fanIn)})
     }